requirements.txt
tensorflow>=1.8.0,<2.0.0
gym[atari,classic_control]>=0.11
scipy
joblib
cloudpickle>=0.5.5
opencv-python
numpy
pandas
matplotlib
pytorch==1.6.0
box2d-py==2.3.5
pybullet
gym-minigrid
scikit-optimize
optuna
pytablewriter
seaborn
pyyaml>=5.1
